refactor : replace Hero component to UpperLayout by using composition pattern
