+++ src/components/hero/Hero/Hero.tsx
@@ -1,35 +1,29 @@
 import React from 'react';
 import styled from '@emotion/styled';
 
-import HeroText from '../HeroText';
-import TagSearch from 'src/components/TagSearch/TagSearch';
-
 interface Props {
   text: string;
   listLength: number;
 }
 
 const Hero = ({ text, listLength }: Props) => {
-  return (
-    <>
-      <Layout>
-        <HeroText text={text} listLength={listLength} />
-        <TagSearch />
-      </Layout>
-      <Hr />
-    </>
-  );
+  return <Text listLength={listLength}>{text}</Text>;
 };
 
+interface StyleProps {
+  listLength: number;
+}
+
+const Text = styled.h1<StyleProps>`
+  font-size: 80px;
+  position: relative;
+  margin-bottom: 30px;
+  :after {
+    content: '${(props) => props.listLength}';
+    font-size: 40px;
+    position: absolute;
+    top: 0;
+  }
+`;
+
 export default Hero;
-
-const Layout = styled.div`
-  width: 1200px;
-  margin: 0px auto;
-  margin-top: 120px;
-  margin-bottom: 50px;
-`;
-
-const Hr = styled.hr`
-  border: 0.5px solid #494c56;
-`;
