+++ src/components/hero/UpperLayout.tsx
